Add Support for Custom CA Certificates in NIMCacheReconciler for Proxy Environments
This PR introduces support for custom CA certificates in the NIMCacheReconciler, enabling secure communication in proxy environments and Kubernetes clusters that require custom certificates. Certificates are managed via a ConfigMap and are automatically mounted into the NIM cache job. This enhancement ensures compatibility with environments that use self-signed or custom CA certificates

Signed-off-by: Shiva Krishna, Merla <smerla@nvidia.com>